Module aws_sdk_elasticloadbalancingv2::client::fluent_builders

Utilities to ergonomically construct a request to the service.

Fluent builders are created through the Client by calling one if its operation methods. After parameters are set using the builder methods, the send method can be called to initiate the request.

Structs

AddListenerCertificates

Fluent builder constructing a request to AddListenerCertificates.

AddTags

Fluent builder constructing a request to AddTags.

CreateListener

Fluent builder constructing a request to CreateListener.

CreateLoadBalancer

Fluent builder constructing a request to CreateLoadBalancer.

CreateRule

Fluent builder constructing a request to CreateRule.

CreateTargetGroup

Fluent builder constructing a request to CreateTargetGroup.

DeleteListener

Fluent builder constructing a request to DeleteListener.

DeleteLoadBalancer

Fluent builder constructing a request to DeleteLoadBalancer.

DeleteRule

Fluent builder constructing a request to DeleteRule.

DeleteTargetGroup

Fluent builder constructing a request to DeleteTargetGroup.

DeregisterTargets

Fluent builder constructing a request to DeregisterTargets.

DescribeAccountLimits

Fluent builder constructing a request to DescribeAccountLimits.

DescribeListenerCertificates

Fluent builder constructing a request to DescribeListenerCertificates.

DescribeListeners

Fluent builder constructing a request to DescribeListeners.

DescribeLoadBalancerAttributes

Fluent builder constructing a request to DescribeLoadBalancerAttributes.

DescribeLoadBalancers

Fluent builder constructing a request to DescribeLoadBalancers.

DescribeRules

Fluent builder constructing a request to DescribeRules.

DescribeSSLPolicies

Fluent builder constructing a request to DescribeSSLPolicies.

DescribeTags

Fluent builder constructing a request to DescribeTags.

DescribeTargetGroupAttributes

Fluent builder constructing a request to DescribeTargetGroupAttributes.

DescribeTargetGroups

Fluent builder constructing a request to DescribeTargetGroups.

DescribeTargetHealth

Fluent builder constructing a request to DescribeTargetHealth.

ModifyListener

Fluent builder constructing a request to ModifyListener.

ModifyLoadBalancerAttributes

Fluent builder constructing a request to ModifyLoadBalancerAttributes.

ModifyRule

Fluent builder constructing a request to ModifyRule.

ModifyTargetGroup

Fluent builder constructing a request to ModifyTargetGroup.

ModifyTargetGroupAttributes

Fluent builder constructing a request to ModifyTargetGroupAttributes.

RegisterTargets

Fluent builder constructing a request to RegisterTargets.

RemoveListenerCertificates

Fluent builder constructing a request to RemoveListenerCertificates.

RemoveTags

Fluent builder constructing a request to RemoveTags.

SetIpAddressType

Fluent builder constructing a request to SetIpAddressType.

SetRulePriorities

Fluent builder constructing a request to SetRulePriorities.

SetSecurityGroups

Fluent builder constructing a request to SetSecurityGroups.

SetSubnets

Fluent builder constructing a request to SetSubnets.