Module aws_sdk_elasticloadbalancingv2::error

Modules

allocation_id_not_found_exception

See AllocationIdNotFoundException

alpn_policy_not_supported_exception

See AlpnPolicyNotSupportedException

availability_zone_not_supported_exception

See AvailabilityZoneNotSupportedException

certificate_not_found_exception

See CertificateNotFoundException

duplicate_listener_exception

See DuplicateListenerException

duplicate_load_balancer_name_exception

See DuplicateLoadBalancerNameException

duplicate_tag_keys_exception

See DuplicateTagKeysException

duplicate_target_group_name_exception

See DuplicateTargetGroupNameException

health_unavailable_exception

See HealthUnavailableException

incompatible_protocols_exception

See IncompatibleProtocolsException

invalid_configuration_request_exception

See InvalidConfigurationRequestException

invalid_load_balancer_action_exception

See InvalidLoadBalancerActionException

invalid_scheme_exception

See InvalidSchemeException

invalid_security_group_exception

See InvalidSecurityGroupException

invalid_subnet_exception

See InvalidSubnetException

invalid_target_exception

See InvalidTargetException

listener_not_found_exception

See ListenerNotFoundException

load_balancer_not_found_exception

See LoadBalancerNotFoundException

operation_not_permitted_exception

See OperationNotPermittedException

priority_in_use_exception

See PriorityInUseException

resource_in_use_exception

See ResourceInUseException

rule_not_found_exception

See RuleNotFoundException

ssl_policy_not_found_exception

See SslPolicyNotFoundException

subnet_not_found_exception

See SubnetNotFoundException

target_group_association_limit_exception

See TargetGroupAssociationLimitException

target_group_not_found_exception

See TargetGroupNotFoundException

too_many_actions_exception

See TooManyActionsException

too_many_certificates_exception

See TooManyCertificatesException

too_many_listeners_exception

See TooManyListenersException

too_many_load_balancers_exception

See TooManyLoadBalancersException

too_many_registrations_for_target_id_exception

See TooManyRegistrationsForTargetIdException

too_many_rules_exception

See TooManyRulesException

too_many_tags_exception

See TooManyTagsException

too_many_target_groups_exception

See TooManyTargetGroupsException

too_many_targets_exception

See TooManyTargetsException

too_many_unique_target_groups_per_load_balancer_exception

See TooManyUniqueTargetGroupsPerLoadBalancerException

unsupported_protocol_exception

See UnsupportedProtocolException

Structs

AddListenerCertificatesError

AddTagsError

AllocationIdNotFoundException

The specified allocation ID does not exist.

AlpnPolicyNotSupportedException

The specified ALPN policy is not supported.

AvailabilityZoneNotSupportedException

The specified Availability Zone is not supported.

CertificateNotFoundException

The specified certificate does not exist.

CreateListenerError

CreateLoadBalancerError

CreateRuleError

CreateTargetGroupError

DeleteListenerError

DeleteLoadBalancerError

DeleteRuleError

DeleteTargetGroupError

DeregisterTargetsError

DescribeAccountLimitsError

DescribeListenerCertificatesError

DescribeListenersError

DescribeLoadBalancerAttributesError

DescribeLoadBalancersError

DescribeRulesError

DescribeSSLPoliciesError

DescribeTagsError

DescribeTargetGroupAttributesError

DescribeTargetGroupsError

DescribeTargetHealthError

DuplicateListenerException

A listener with the specified port already exists.

DuplicateLoadBalancerNameException

A load balancer with the specified name already exists.

DuplicateTagKeysException

A tag key was specified more than once.

DuplicateTargetGroupNameException

A target group with the specified name already exists.

HealthUnavailableException

The health of the specified targets could not be retrieved due to an internal

IncompatibleProtocolsException

The specified configuration is not valid with this protocol.

InvalidConfigurationRequestException

The requested configuration is not valid.

InvalidLoadBalancerActionException

The requested action is not valid.

InvalidSchemeException

The requested scheme is not valid.

InvalidSecurityGroupException

The specified security group does not exist.

InvalidSubnetException

The specified subnet is out of available addresses.

InvalidTargetException

The specified target does not exist, is not in the same VPC as the target group, or has an

ListenerNotFoundException

The specified listener does not exist.

LoadBalancerNotFoundException

The specified load balancer does not exist.

ModifyListenerError

ModifyLoadBalancerAttributesError

ModifyRuleError

ModifyTargetGroupAttributesError

ModifyTargetGroupError

OperationNotPermittedException

This operation is not allowed.

PriorityInUseException

The specified priority is in use.

RegisterTargetsError

RemoveListenerCertificatesError

RemoveTagsError

ResourceInUseException

A specified resource is in use.

RuleNotFoundException

The specified rule does not exist.

SetIpAddressTypeError

SetRulePrioritiesError

SetSecurityGroupsError

SetSubnetsError

SslPolicyNotFoundException

The specified SSL policy does not exist.

SubnetNotFoundException

The specified subnet does not exist.

TargetGroupAssociationLimitException

You've reached the limit on the number of load balancers per target group.

TargetGroupNotFoundException

The specified target group does not exist.

TooManyActionsException

You've reached the limit on the number of actions per rule.

TooManyCertificatesException

You've reached the limit on the number of certificates per load balancer.

TooManyListenersException

You've reached the limit on the number of listeners per load balancer.

TooManyLoadBalancersException

You've reached the limit on the number of load balancers for your Amazon Web Services

TooManyRegistrationsForTargetIdException

You've reached the limit on the number of times a target can be registered with a load

TooManyRulesException

You've reached the limit on the number of rules per load balancer.

TooManyTagsException

You've reached the limit on the number of tags per load balancer.

TooManyTargetGroupsException

You've reached the limit on the number of target groups for your Amazon Web Services

TooManyTargetsException

You've reached the limit on the number of targets.

TooManyUniqueTargetGroupsPerLoadBalancerException

You've reached the limit on the number of unique target groups per load balancer across

UnsupportedProtocolException

The specified protocol is not supported.

Enums

AddListenerCertificatesErrorKind

AddTagsErrorKind

CreateListenerErrorKind

CreateLoadBalancerErrorKind

CreateRuleErrorKind

CreateTargetGroupErrorKind

DeleteListenerErrorKind

DeleteLoadBalancerErrorKind

DeleteRuleErrorKind

DeleteTargetGroupErrorKind

DeregisterTargetsErrorKind

DescribeAccountLimitsErrorKind

DescribeListenerCertificatesErrorKind

DescribeListenersErrorKind

DescribeLoadBalancerAttributesErrorKind

DescribeLoadBalancersErrorKind

DescribeRulesErrorKind

DescribeSSLPoliciesErrorKind

DescribeTagsErrorKind

DescribeTargetGroupAttributesErrorKind

DescribeTargetGroupsErrorKind

DescribeTargetHealthErrorKind

ModifyListenerErrorKind

ModifyLoadBalancerAttributesErrorKind

ModifyRuleErrorKind

ModifyTargetGroupAttributesErrorKind

ModifyTargetGroupErrorKind

RegisterTargetsErrorKind

RemoveListenerCertificatesErrorKind

RemoveTagsErrorKind

SetIpAddressTypeErrorKind

SetRulePrioritiesErrorKind

SetSecurityGroupsErrorKind

SetSubnetsErrorKind