Module aws_sdk_elasticloadbalancingv2::model

Modules

action

See Action

authenticate_cognito_action_config

See AuthenticateCognitoActionConfig

authenticate_oidc_action_config

See AuthenticateOidcActionConfig

availability_zone

See AvailabilityZone

certificate

See Certificate

cipher

See Cipher

fixed_response_action_config

See FixedResponseActionConfig

forward_action_config

See ForwardActionConfig

host_header_condition_config

See HostHeaderConditionConfig

http_header_condition_config

See HttpHeaderConditionConfig

http_request_method_condition_config

See HttpRequestMethodConditionConfig

limit

See Limit

listener

See Listener

load_balancer

See LoadBalancer

load_balancer_address

See LoadBalancerAddress

load_balancer_attribute

See LoadBalancerAttribute

load_balancer_state

See LoadBalancerState

matcher

See Matcher

path_pattern_condition_config

See PathPatternConditionConfig

query_string_condition_config

See QueryStringConditionConfig

query_string_key_value_pair

See QueryStringKeyValuePair

redirect_action_config

See RedirectActionConfig

rule

See Rule

rule_condition

See RuleCondition

rule_priority_pair

See RulePriorityPair

source_ip_condition_config

See SourceIpConditionConfig

ssl_policy

See SslPolicy

subnet_mapping

See SubnetMapping

tag

See Tag

tag_description

See TagDescription

target_description

See TargetDescription

target_group

See TargetGroup

target_group_attribute

See TargetGroupAttribute

target_group_stickiness_config

See TargetGroupStickinessConfig

target_group_tuple

See TargetGroupTuple

target_health

See TargetHealth

target_health_description

See TargetHealthDescription

Structs

Action

Information about an action.

AuthenticateCognitoActionConfig

Request parameters to use when integrating with Amazon Cognito to authenticate

AuthenticateOidcActionConfig

Request parameters when using an identity provider (IdP) that is compliant with OpenID

AvailabilityZone

Information about an Availability Zone.

Certificate

Information about an SSL server certificate.

Cipher

Information about a cipher used in a policy.

FixedResponseActionConfig

Information about an action that returns a custom HTTP response.

ForwardActionConfig

Information about a forward action.

HostHeaderConditionConfig

Information about a host header condition.

HttpHeaderConditionConfig

Information about an HTTP header condition.

HttpRequestMethodConditionConfig

Information about an HTTP method condition.

Limit

Information about an Elastic Load Balancing resource limit for your Amazon Web Services

Listener

Information about a listener.

LoadBalancer

Information about a load balancer.

LoadBalancerAddress

Information about a static IP address for a load balancer.

LoadBalancerAttribute

Information about a load balancer attribute.

LoadBalancerState

Information about the state of the load balancer.

Matcher

The codes to use when checking for a successful response from a target. If the protocol

PathPatternConditionConfig

Information about a path pattern condition.

QueryStringConditionConfig

Information about a query string condition.

QueryStringKeyValuePair

Information about a key/value pair.

RedirectActionConfig

Information about a redirect action.

Rule

Information about a rule.

RuleCondition

Information about a condition for a rule.

RulePriorityPair

Information about the priorities for the rules for a listener.

SourceIpConditionConfig

Information about a source IP condition.

SslPolicy

Information about a policy used for SSL negotiation.

SubnetMapping

Information about a subnet mapping.

Tag

Information about a tag.

TagDescription

The tags associated with a resource.

TargetDescription

Information about a target.

TargetGroup

Information about a target group.

TargetGroupAttribute

Information about a target group attribute.

TargetGroupStickinessConfig

Information about the target group stickiness for a rule.

TargetGroupTuple

Information about how traffic will be distributed between multiple target groups in a

TargetHealth

Information about the current health of a target.

TargetHealthDescription

Information about the health of a target.

Enums

ActionTypeEnum

AuthenticateCognitoActionConditionalBehaviorEnum

AuthenticateOidcActionConditionalBehaviorEnum

IpAddressType

LoadBalancerSchemeEnum

LoadBalancerStateEnum

LoadBalancerTypeEnum

ProtocolEnum

RedirectActionStatusCodeEnum

TargetHealthReasonEnum

TargetHealthStateEnum

TargetTypeEnum