Struct aws_sdk_elasticloadbalancingv2::model::load_balancer::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for LoadBalancer

Implementations

impl Builder

pub fn load_balancer_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the load balancer.

pub fn set_load_balancer_arn(self, input: Option<String>) -> Self

pub fn dns_name(self, input: impl Into<String>) -> Self

The public DNS name of the load balancer.

pub fn set_dns_name(self, input: Option<String>) -> Self

pub fn canonical_hosted_zone_id(self, input: impl Into<String>) -> Self

The ID of the Amazon Route 53 hosted zone associated with the load balancer.

pub fn set_canonical_hosted_zone_id(self, input: Option<String>) -> Self

pub fn created_time(self, input: Instant) -> Self

The date and time the load balancer was created.

pub fn set_created_time(self, input: Option<Instant>) -> Self

pub fn load_balancer_name(self, input: impl Into<String>) -> Self

The name of the load balancer.

pub fn set_load_balancer_name(self, input: Option<String>) -> Self

pub fn scheme(self, input: LoadBalancerSchemeEnum) -> Self

The nodes of an Internet-facing load balancer have public IP addresses. The DNS name of an Internet-facing load balancer is publicly resolvable to the public IP addresses of the nodes. Therefore, Internet-facing load balancers can route requests from clients over the internet.

The nodes of an internal load balancer have only private IP addresses. The DNS name of an internal load balancer is publicly resolvable to the private IP addresses of the nodes. Therefore, internal load balancers can route requests only from clients with access to the VPC for the load balancer.

pub fn set_scheme(self, input: Option<LoadBalancerSchemeEnum>) -> Self

pub fn vpc_id(self, input: impl Into<String>) -> Self

The ID of the VPC for the load balancer.

pub fn set_vpc_id(self, input: Option<String>) -> Self

pub fn state(self, input: LoadBalancerState) -> Self

The state of the load balancer.

pub fn set_state(self, input: Option<LoadBalancerState>) -> Self

pub fn type(self, input: LoadBalancerTypeEnum) -> Self

The type of load balancer.

pub fn set_type(self, input: Option<LoadBalancerTypeEnum>) -> Self

pub fn availability_zones(self, input: impl Into<AvailabilityZone>) -> Self

pub fn set_availability_zones(
self,
input: Option<Vec<AvailabilityZone>>
) -> Self

pub fn security_groups(self, input: impl Into<String>) -> Self

pub fn set_security_groups(self, input: Option<Vec<String>>) -> Self

pub fn ip_address_type(self, input: IpAddressType) -> Self

The type of IP addresses used by the subnets for your load balancer. The possible values are ipv4 (for IPv4 addresses) and dualstack (for IPv4 and IPv6 addresses).

pub fn set_ip_address_type(self, input: Option<IpAddressType>) -> Self

pub fn customer_owned_ipv4_pool(self, input: impl Into<String>) -> Self

[Application Load Balancers on Outposts] The ID of the customer-owned address pool.

pub fn set_customer_owned_ipv4_pool(self, input: Option<String>) -> Self

pub fn build(self) -> LoadBalancer

Consumes the builder and constructs a LoadBalancer