Module aws_sdk_elasticloadbalancingv2::operation

Structs

AddListenerCertificates

Adds the specified SSL server certificate to the certificate list for the specified HTTPS

AddTags

Adds the specified tags to the specified Elastic Load Balancing resource. You can tag your

CreateListener

Creates a listener for the specified Application Load Balancer, Network Load Balancer, or

CreateLoadBalancer

Creates an Application Load Balancer, Network Load Balancer, or Gateway Load

CreateRule

Creates a rule for the specified listener. The listener must be associated with an

CreateTargetGroup

Creates a target group.

DeleteListener

Deletes the specified listener.

DeleteLoadBalancer

Deletes the specified Application Load Balancer, Network Load Balancer, or Gateway Load

DeleteRule

Deletes the specified rule.

DeleteTargetGroup

Deletes the specified target group.

DeregisterTargets

Deregisters the specified targets from the specified target group. After the targets are

DescribeAccountLimits

Describes the current Elastic Load Balancing resource limits for your Amazon Web Services

DescribeListenerCertificates

Describes the default certificate and the certificate list for the specified HTTPS or TLS

DescribeListeners

Describes the specified listeners or the listeners for the specified Application Load

DescribeLoadBalancerAttributes

Describes the attributes for the specified Application Load Balancer, Network Load

DescribeLoadBalancers

Describes the specified load balancers or all of your load balancers.

DescribeRules

Describes the specified rules or the rules for the specified listener. You must specify

DescribeSSLPolicies

Describes the specified policies or all policies used for SSL negotiation.

DescribeTags

Describes the tags for the specified Elastic Load Balancing resources. You can describe

DescribeTargetGroupAttributes

Describes the attributes for the specified target group.

DescribeTargetGroups

Describes the specified target groups or all of your target groups. By default, all target

DescribeTargetHealth

Describes the health of the specified targets or all of your targets.

ModifyListener

Replaces the specified properties of the specified listener. Any properties that you do

ModifyLoadBalancerAttributes

Modifies the specified attributes of the specified Application Load Balancer, Network Load

ModifyRule

Replaces the specified properties of the specified rule. Any properties that you do not

ModifyTargetGroup

Modifies the health checks used when evaluating the health state of the targets in the

ModifyTargetGroupAttributes

Modifies the specified attributes of the specified target group.

RegisterTargets

Registers the specified targets with the specified target group.

RemoveListenerCertificates

Removes the specified certificate from the certificate list for the specified HTTPS or TLS

RemoveTags

Removes the specified tags from the specified Elastic Load Balancing resources. You can

SetIpAddressType

Sets the type of IP addresses used by the subnets of the specified Application Load

SetRulePriorities

Sets the priorities of the specified rules.

SetSecurityGroups

Associates the specified security groups with the specified Application Load Balancer. The

SetSubnets

Enables the Availability Zones for the specified public subnets for the specified